+++ constants.py
@@ -47,3 +47,30 @@ STATES = [
     'Wisconsin',
     'Wyoming'
 ]
+
+HOUSING_STOCK_DISTRIBUTION = {
+    # Assumed National housing distribution [https://www.census.gov/programs-surveys/ahs/data/interactive/ahstablecreator.html?s_areas=00000&s_year=2017&s_tablename=TABLE2&s_bygroup1=1&s_bygroup2=1&s_filtergroup1=1&s_filtergroup2=1]
+    0: 0.0079,
+    1: 0.1083,
+    2: 0.2466,
+    3: 0.4083,
+    4: 0.2289
+}
+
+BURDENED_HOUSEHOLD_PROPORTION = [5, 25, 33, 50, 75]
+
+COLOR_RANGE = [
+    [65, 182, 196],
+    [127, 205, 187],
+    [199, 233, 180],
+    [237, 248, 177],
+    [255, 255, 204],
+    [255, 237, 160],
+    [254, 217, 118],
+    [254, 178, 76],
+    [253, 141, 60],
+    [252, 78, 42],
+    [227, 26, 28],
+]
+
+BREAKS = [0, .1, .2, .3, .4, .5, .6, .7, .8, .9, 1]
