+++ requirements.txt
@@ -73,7 +73,6 @@ pyarrow==1.0.1
 pycparser==2.20
 pydeck==0.5.0b1
 Pygments==2.7.1
-pyinstaller==4.0
 pyinstaller-hooks-contrib==2020.7
 pyparsing==2.4.7
 pyproj==2.6.1.post1
@@ -93,7 +92,7 @@ six==1.15.0
 sklearn==0.0
 smmap==3.0.4
 SQLAlchemy==1.3.20
-streamlit==0.70.0
+streamlit==0.72.0
 terminado==0.9.1
 testpath==0.4.4
 threadpoolctl==2.1.0
