+++ utils.py
@@ -1,0 +1,27 @@
+import base64
+import pandas as pd
+from six import BytesIO
+
+
+def to_excel(df: pd.DataFrame):
+    output = BytesIO()
+    writer = pd.ExcelWriter(output, engine='xlsxwriter')
+    df.to_excel(writer, sheet_name='Sheet1')
+    writer.save()
+    processed_data = output.getvalue()
+    return processed_data
+
+
+def get_table_download_link(df: pd.DataFrame, file_name: str, text: str):
+    """Generates a link allowing the data in a given panda dataframe to be downloaded
+    in:  dataframe
+    out: href string
+    """
+    val = to_excel(df)
+    b64 = base64.b64encode(val)  # val looks like b'...'
+    return f'<a href="data:application/octet-stream;base64,{b64.decode()}" download="{file_name}.xlsx">{text}</a>'
+
+
+def output_table(df: pd.DataFrame, path: str):
+    df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
+    df.to_excel(path)
